[feature/popular-articles]: [SSR] [Refactoring] Update imports
